+++ packages/adapter-mongoose/lib/adapter-mongoose.js
@@ -14,61 +14,16 @@ const {
   BaseListAdapter,
   BaseFieldAdapter,
 } = require('@keystone-alpha/keystone');
-const { mongoJoinBuilder } = require('@keystone-alpha/mongo-join-builder');
+const {
+  queryParser,
+  pipelineBuilder,
+  mutationBuilder,
+} = require('@keystone-alpha/mongo-join-builder');
 const logger = require('@keystone-alpha/logger').logger('mongoose');
 
-const simpleTokenizer = require('./tokenizers/simple');
-const relationshipTokenizer = require('./tokenizers/relationship');
-const getRelatedListAdapterFromQueryPathFactory = require('./tokenizers/relationship-path');
 const slugify = require('@sindresorhus/slugify');
 
 const debugMongoose = () => !!process.env.DEBUG_MONGOOSE;
-
-const modifierConditions = {
-  // TODO: Implement configurable search fields for lists
-  $search: value => {
-    if (!value || (getType(value) === 'String' && !value.trim())) {
-      return undefined;
-    }
-    return {
-      $match: {
-        name: new RegExp(`${escapeRegExp(value)}`, 'i'),
-      },
-    };
-  },
-
-  $orderBy: (value, _, listAdapter) => {
-    const [orderField, orderDirection] = value.split('_');
-
-    const mongoField = listAdapter.graphQlQueryPathToMongoField(orderField);
-
-    return {
-      $sort: {
-        [mongoField]: orderDirection === 'DESC' ? -1 : 1,
-      },
-    };
-  },
-
-  $skip: value => {
-    if (value < Infinity && value > 0) {
-      return {
-        $skip: value,
-      };
-    }
-  },
-
-  $first: value => {
-    if (value < Infinity && value > 0) {
-      return {
-        $limit: value,
-      };
-    }
-  },
-
-  $count: value => ({
-    $count: value,
-  }),
-};
 
 class MongooseAdapter extends BaseKeystoneAdapter {
   constructor() {
@@ -162,20 +117,6 @@ class MongooseListAdapter extends BaseListAdapter {
 
     // Need to call postConnect() once all fields have registered and the database is connected to.
     this.model = null;
-
-    this.queryBuilder = mongoJoinBuilder({
-      tokenizer: {
-        // executed for simple query components (eg; 'fulfilled: false' / name: 'a')
-        simple: simpleTokenizer({
-          getRelatedListAdapterFromQueryPath: getRelatedListAdapterFromQueryPathFactory(this),
-          modifierConditions,
-        }),
-        // executed for complex query components (eg; items: { ... })
-        relationship: relationshipTokenizer({
-          getRelatedListAdapterFromQueryPath: getRelatedListAdapterFromQueryPathFactory(this),
-        }),
-      },
-    });
   }
 
   prepareFieldAdapter(fieldAdapter) {
@@ -297,8 +238,14 @@ class MongooseListAdapter extends BaseListAdapter {
       query.$count = 'count';
     }
 
-    return this.queryBuilder(query, pipeline => this.model.aggregate(pipeline).exec()).then(
-      foundItems => {
+    const queryTree = queryParser({ listAdapter: this }, query);
+
+    // Run the query against the given database and collection
+    return this.model
+      .aggregate(pipelineBuilder(queryTree))
+      .exec()
+      .then(mutationBuilder(queryTree.relationships))
+      .then(foundItems => {
         if (meta) {
           // When there are no items, we get undefined back, so we simulate the
           // normal result of 0 items.
@@ -308,8 +255,7 @@ class MongooseListAdapter extends BaseListAdapter {
           return foundItems[0];
         }
         return foundItems;
-      }
-    );
+      });
   }
 }