+++ packages/fields/src/types/Relationship/Implementation.js
@@ -1,5 +1,4 @@
 import mongoose from 'mongoose';
-import omitBy from 'lodash.omitby';
 import { mergeWhereClause } from '@keystone-alpha/utils';
 import { MongooseFieldAdapter } from '@keystone-alpha/adapter-mongoose';
 import { KnexFieldAdapter } from '@keystone-alpha/adapter-knex';
@@ -110,7 +109,7 @@ export class Relationship extends Implementation {
     // to-one relationships are much easier to deal with.
     if (!this.many) {
       return {
-        [this.path]: (item, _, context) => {
+        [this.path]: (item, _, context, info) => {
           // No ID set, so we return null for the value
           if (!item[this.path]) {
             return null;
@@ -118,7 +117,7 @@ export class Relationship extends Implementation {
           const filteredQueryArgs = { where: { id: item[this.path].toString() } };
           // We do a full query to ensure things like access control are applied
           return refList
-            .listQuery(filteredQueryArgs, context, refList.gqlNames.listQueryName)
+            .listQuery(filteredQueryArgs, context, refList.gqlNames.listQueryName, info)
             .then(items => (items && items.length ? items[0] : null));
         },
       };
@@ -143,15 +142,15 @@ export class Relationship extends Implementation {
     };
 
     return {
-      [this.path]: (item, args, context, { fieldName }) => {
+      [this.path]: (item, args, context, info) => {
         const filteredQueryArgs = buildManyQueryArgs(item, args);
-        return refList.listQuery(filteredQueryArgs, context, fieldName);
+        return refList.listQuery(filteredQueryArgs, context, info.fieldName, info);
       },
 
       ...(this.withMeta && {
-        [`_${this.path}Meta`]: (item, args, context, { fieldName }) => {
+        [`_${this.path}Meta`]: (item, args, context, info) => {
           const filteredQueryArgs = buildManyQueryArgs(item, args);
-          return refList.listQueryMeta(filteredQueryArgs, context, fieldName);
+          return refList.listQueryMeta(filteredQueryArgs, context, info.fieldName, info);
         },
       }),
     };
@@ -394,51 +393,6 @@ export class MongoRelationshipInterface extends MongooseFieldAdapter {
       [`${this.path}_is_null`]: value => ({
         [dbPath]: value ? { $not: { $exists: true, $ne: null } } : { $exists: true, $ne: null },
       }),
-    };
-  }
-
-  getRelationshipQueryCondition(queryKey, uid) {
-    const filterType = {
-      [this.path]: 'every',
-      [`${this.path}_every`]: 'every',
-      [`${this.path}_some`]: 'some',
-      [`${this.path}_none`]: 'none',
-    }[queryKey];
-
-    return {
-      from: this.getRefListAdapter().model.collection.name, // the collection name to join with
-      field: this.path, // The field on this collection
-      // A mutation to run on the data post-join. Useful for merging joined
-      // data back into the original object.
-      // Executed on a depth-first basis for nested relationships.
-      postQueryMutation: (parentObj /*, keyOfRelationship, rootObj, pathToParent*/) => {
-        return omitBy(
-          parentObj,
-          /*
-          {
-            ...parentObj,
-            // Given there could be sorting and limiting that's taken place, we
-            // want to overwrite the entire object rather than merging found items
-            // in.
-            [field]: parentObj[keyOfRelationship],
-          },
-          */
-          // Clean up the result to remove the intermediate results
-          (_, keyToOmit) => keyToOmit.startsWith(uid)
-        );
-      },
-      // The conditions under which an item from the 'orders' collection is
-      // considered a match and included in the end result
-      // All the keys on an 'order' are available, plus 3 special keys:
-      // 1) <uid>_<field>_every - is `true` when every joined item matches the
-      //    query
-      // 2) <uid>_<field>_some - is `true` when some joined item matches the
-      //    query
-      // 3) <uid>_<field>_none - is `true` when none of the joined items match
-      //    the query
-      matchTerm: { [`${uid}_${this.path}_${filterType}`]: true },
-      // Flag this is a to-many relationship
-      many: this.field.many,
     };
   }