+++ packages/mongo-join-builder/index.js
@@ -1,3 +1,8 @@
-const mongoJoinBuilder = require('./lib/index');
+const { queryParser } = require('./lib/query-parser');
+const { pipelineBuilder, mutationBuilder } = require('./lib/join-builder');
 
-module.exports = { mongoJoinBuilder };
+module.exports = {
+  queryParser,
+  pipelineBuilder,
+  mutationBuilder,
+};