+++ demo-projects/todo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @keystone-alpha/demo-project-todo
 
+## 1.1.16
+
+- Updated dependencies [4e6a574d](https://github.com/keystonejs/keystone-5/commit/4e6a574d):
+- Updated dependencies [b96a3a58](https://github.com/keystonejs/keystone-5/commit/b96a3a58):
+  - @keystone-alpha/app-admin-ui@5.10.0
+  - @keystone-alpha/keystone@15.3.1
+  - @keystone-alpha/fields@13.0.0
+  - @keystone-alpha/adapter-mongoose@5.0.0
+
+## 1.1.15
+
+### Patch Changes
+
+- [9b532072](https://github.com/keystonejs/keystone-5/commit/9b532072): Rename Keystone to KeystoneJS in docs where possible in docs
+
 ## 1.1.14
 
 - Updated dependencies [42a45bbd](https://github.com/keystonejs/keystone-5/commit/42a45bbd):